requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile /workspace/requirements.in -o /workspace/requirements.txt
amqp==5.3.1
    # via kombu
asgiref==3.10.0
    # via django
billiard==4.2.2
    # via celery
celery==5.5.3
    # via -r requirements.in
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
django==5.2.8
    # via
    #   -r requirements.in
    #   django-redis
    #   djangorestframework
django-redis==5.4.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
kombu==5.5.4
    # via celery
packaging==25.0
    # via
    #   kombu
    #   pytest
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via click-repl
psycopg==3.2.12
    # via -r requirements.in
psycopg-binary==3.2.12
    # via psycopg
pytest==8.3.4
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.9.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
redis==5.2.1
    # via
    #   -r requirements.in
    #   django-redis
    #   kombu
requests==2.32.5
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via kombu
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
